feat(Aggegare parameter selection): Adding support of selection of all parameter selections
